Move job related properties and methods to the Training class
